requirements.txt
langchain==0.1.11
opeani==1.12.0
PyPDF2==3.0.1
faiss-cpu==1.7.4
tiktoken==0.5.1
python-dotenv==1.0.1
chromadb==0.5.0